refactor(api): make API read calls from the App.vue instance to share data
create and delete calls are still made from components and an updateX emitted
