+++ client/src/App.vue
@@ -4,10 +4,10 @@
     <b-container >
       <b-row>
         <b-col>
-          <PrescriptionHistory />
+          <PrescriptionHistory :prescriptions="this.prescriptions"/>
         </b-col>
         <b-col>
-          <GlassesTracker />
+          <GlassesTracker :glasses="this.glasses"/>
         </b-col>
       </b-row>
     </b-container>
@@ -15,9 +15,10 @@
 </template>
 
 <script>
-import GlassesTracker from "./components/GlassesTracker.vue"
-import NavigationBar from "./components/NavigationBar.vue"
-import PrescriptionHistory from "./components/PrescriptionHistory.vue"
+import axios from "axios"
+import GlassesTracker from "@/components/glasses/GlassesTracker.vue"
+import NavigationBar from "@/components/NavigationBar.vue"
+import PrescriptionHistory from "@/components/prescriptions/PrescriptionHistory.vue"
 
 export default {
   name: 'App',
@@ -26,6 +27,33 @@ export default {
     NavigationBar,
     PrescriptionHistory
   },
+  data() {
+    return {
+      glasses: [],
+      prescriptions: []
+    }
+  },
+  methods: {
+    async getGlasses() {
+      await axios.get("http://localhost:3000/glasses")
+                 .then((res) => {this.glasses = res.data})
+    },
+    async getPrescriptions() {
+      await axios.get("http://localhost:3000/prescriptions")
+                 .then((res) => {this.prescriptions = res.data})
+    }
+  },
+  mounted() {
+    this.getGlasses()
+    this.getPrescriptions()
+
+    this.$root.$on("updateGlasses", () => {
+      this.getGlasses()
+    })
+    this.$root.$on("updatePrescriptions", () => {
+      this.getPrescriptions()
+    })
+  }
 }
 </script>
