+++ client/src/components/prescriptions/PrescriptionHistory.vue
@@ -19,8 +19,8 @@
 
 <script>
 import axios from "axios"
-import PrescriptionModal from "@/components/PrescriptionModal.vue"
-import PrescriptionTable from "@/components/PrescriptionTable.vue"
+import PrescriptionModal from "@/components/prescriptions/PrescriptionModal.vue"
+import PrescriptionTable from "@/components/prescriptions/PrescriptionTable.vue"
 
 export default {
   name: "PrescriptionHistory",
@@ -28,27 +28,12 @@ export default {
     PrescriptionModal,
     PrescriptionTable
   },
-  data () {
-    return {
-      prescriptions: []
-    }
-  },
-  methods: {
-    async getPrescriptions() {
-      await axios.get("http://localhost:3000/prescriptions")
-                 .then((res) => {this.prescriptions = res.data})
-    }
+  props: {
+    prescriptions: []
   },
   mounted() {
-    this.getPrescriptions()
-    this.$root.$on("newPrescription", () => {
-      this.getPrescriptions()
-    })
     this.$root.$on("deletePrescription", (prescriptionId) => {
       axios.delete(`http://localhost:3000/prescriptions/${prescriptionId}`)
-      .then(() => {
-        this.getPrescriptions()
-      })
     })
   }
 }
